components/asp/requirements.txt
mysql-connector-python
clp-ffi-py